requirements.txt
colorama==0.4.6
pre_commit==4.2.0
python-dotenv==1.1.0
